+++ authpass/CHANGELOG.md
@@ -1,5 +1,7 @@
-# 1.6.5
+# 1.7.0
 
+* Windows: Initial Windows Support 🎉️
+* Improved group filter by showing recursive filter and trash.
 * WebDav: Connect UI: Make dialog scrollable. #103
 * WebDav: Fixed Compatibility with servers returning trailing slashes for directories.
 * Linux: support for keyring storage (for Dropbox, Google Drive, WebDav, One Drive)